+++ src/generator/settings/module_generator_settings.cpp
@@ -15,7 +15,7 @@
 #include <QFont>
 
 const QString
-ModuleGeneratorSettings::S_VERSION(QStringLiteral("1.1.1"));
+ModuleGeneratorSettings::S_VERSION(QStringLiteral("1.1.2"));
 const QRegularExpression
 ModuleGeneratorSettings::REG_PREFIX(QStringLiteral("[A-Za-z]([A-Za-z]|\\d)*"));
 const QRegularExpression