+++ src/widgets/class_specification/class_specification_widget.cpp
@@ -12,6 +12,7 @@
 #include <QRegularExpressionValidator>
 #include <QCheckBox>
 #include <QPushButton>
+#include <QCloseEvent>
 
 //const static int I_MINIMAL_WIDTH = 320;