+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Changed
+- supports version GTlab 2.0 alpha - #31
+
+## Added
+- Compatibility file - #31
 
 ## [1.1.1] - 2022-07-25
 ## Changed